requirements.txt
tensorflow==2.16.1
tensorflow-model-optimization==0.8.0
tf-keras
numpy==1.26.4
pandas==2.2.1
matplotlib
seaborn